Extract Number enum to its own module
Moves the Number enum from the number_property module to its own number module within the domain.

This change improves code organization and reduces redundancy as the Number enum is used in multiple modules.
