+++ src/domain/mod.rs
@@ -3,4 +3,7 @@ pub mod controller;
 pub mod controller_registry;
 pub mod device;
 pub mod events;
+mod number;
 pub mod property;
+
+pub use number::Number;
